requirements.txt
# Axiom Engine - Core Dependencies
# --------------------------------
# To install all dependencies for this project, including the AI model,
# simply run the following command from the project's root directory:
#
# pip3 install -r requirements.txt
#

# --- Core Web & Networking ---
Flask==3.0.3
gunicorn==22.0.0
requests==2.31.0

# --- Web Scraping & Data Extraction ---
googlesearch-python==1.2.4
trafilatura==1.9.0
beautifulsoup4==4.12.3

# --- API Interaction ---
newsapi-python==0.2.7
serpapi==2.1.0


# --- AI & Natural Language Processing ---
spacy==3.7.4

# Direct link to the required spaCy model.
# This allows for a fully automated, one-step installation.
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
